backend/requirements.txt
# Django Core
Django>=5.0,<5.2
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5

# Database
dj-database-url>=2.1

# ML / Inference
ultralytics>=8.3.0
Pillow>=10.2

# Utilities
python-dotenv>=1.0
gunicorn>=21.2
whitenoise>=6.6
